- Support : Add Page to key in serial number - Support : Adjust Leyout, Add Sale order detail - Support : Edit DataModule.cs for updating image to replace none image background in commerce-support.cshtml - Remove content part in laptop-detail page (product type laptop) by editing in NancyBlackRazorViewBase.cs - Add WaitingForOrder in SaleOrderStatus class in SaleOrder.cs - Fixed: wrong billing address in commerce-print.cshtml - Add : Support link button and Add button for editing product in SaleOrder - Removed: CommerceModule.SetN15Price function in CommerceModule.cs (this function used for automate price reversion from promotion to normal price) - Add : Link to WaitingForOrder in _admin.cshtml

+++ Modules/DatabaseSystem/DataModule.cs
@@ -320,7 +320,12 @@ namespace NantCom.NancyBlack.Modules
                             // delete the one being replaced
                             var directory = this.GetAttachmentFolder(tableName, id);
                             var toDelete = Path.Combine(directory, Path.GetFileName( item["Url"].ToString() ));
-                            File.Delete(toDelete);
+
+                            // delete when url is not null
+                            if (!string.IsNullOrEmpty(item["Url"].ToString()))
+                            {
+                                File.Delete(toDelete);
+                            }
                             BaseDataModule.AttachmentDeleted(this.Context, tableName, contentItem, item["Url"].ToString());
 
                             // set to new one
